Move app files to create app folder
